Move Help and About to components/guidance-content
